requirements.txt
confluent-kafka
apache-flink
boto3
pymongo
dotenv
python-dotenv
apache-airflow
psycopg2-binary
pandas